+++ releaseNotes.md
@@ -1,3 +1,14 @@
+_June 10th, 2019_  **v2.0.33**
+- PostLocationHeaderResultFilter - Changed Critical message to warning.
+- APIBlox.AspNetCore.CommandsQueriesControllersOhMy.Templates.Post - Changed fail message to warning.
+  
+Thanks,  
+_Slacquer_  
+
+<br>
+<br>
+
+
 _June 10th, 2019_  **v2.0.32**
 - IReadOnlyEventStoreService.ReadEventStreamVersionAsync, will now return NULL when a stream is not found.
 - ComposedTemplatesExtensions.WriteQueryByController, was too restrictive, as it required an actionRoute with tokens.